+++ src/toc/toc.interface.ts
@@ -4,7 +4,6 @@
 export interface IToc {
   createToc(): string;
   insertToc(): void;
-  setMaxDepth(maxDepth: number): void;
 }
 
 /**